refactor product detail into subcomponents
